+++ rustrade-data/src/exchange/massive/options.rs
@@ -974,9 +974,6 @@ impl MassiveRestClient {
                 }
             }
 
-            if let Some(ref url) = parsed.next_url {
-                Self::validate_next_url(url, base_url)?;
-            }
             next_url = parsed.next_url;
         }
 
@@ -1054,9 +1051,6 @@ impl MassiveRestClient {
                 }
             }
 
-            if let Some(ref url) = parsed.next_url {
-                Self::validate_next_url(url, base_url)?;
-            }
             next_url = parsed.next_url;
         }